requirements.txt
certifi==2019.6.16
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.9.0
mysqlclient==1.3.14
pytz==2019.1
sqlparse==0.3.0
